# Conflicts:
#	app/Console/Commands/SendExpirationAlerts.php
#	composer.lock
#	config/version.php
#	public/css/build/app.css
#	public/css/build/app.css.map
#	public/css/build/overrides.css
#	public/css/build/overrides.css.map
#	public/css/dist/all.css
#	resources/views/users/view.blade.php

+++ app/Http/Controllers/Users/UsersController.php
@@ -57,7 +57,7 @@ class UsersController extends Controller
     public function create(Request $request)
     {
         $this->authorize('create', User::class);
-        $groups = Group::pluck('name', 'id');
+        $groups = Group::orderBy('name', 'asc')->pluck('name', 'id');
 
         $userGroups = collect();
 
@@ -131,10 +131,10 @@ class UsersController extends Controller
         // we have to invoke the form request here to handle image uploads
         app(ImageUploadRequest::class)->handleImages($user, 600, 'avatar', 'avatars', 'avatar');
 
-        if ($request->get('redirect_option') === 'back'){
+        if ($request->input('redirect_option') === 'back'){
             session()->put(['redirect_option' => 'index']);
         } else {
-            session()->put(['redirect_option' => $request->get('redirect_option')]);
+            session()->put(['redirect_option' => $request->input('redirect_option')]);
         }
 
 
@@ -152,7 +152,9 @@ class UsersController extends Controller
             }
 
             if ($request->filled('groups')) {
-                $user->groups()->sync($request->input('groups'));
+                if (auth()->user()->can('canEditAuthFields', $user) && auth()->user()->can('editableOnDemo')) {
+                    $user->groups()->sync($request->input('groups'));
+                }
             } else {
                 $user->groups()->sync([]);
             }
@@ -200,7 +202,7 @@ class UsersController extends Controller
             }
 
             $permissions = config('permissions');
-            $groups = Group::pluck('name', 'id');
+            $groups = Group::orderBy('name', 'asc')->pluck('name', 'id');
 
             $userGroups = $user->groups()->pluck('name', 'id');
             $user->permissions = $user->decodePermissions();
@@ -326,7 +328,7 @@ class UsersController extends Controller
 
         // Handle uploaded avatar
         app(ImageUploadRequest::class)->handleImages($user, 600, 'avatar', 'avatars', 'avatar');
-        session()->put(['redirect_option' => $request->get('redirect_option')]);
+        session()->put(['redirect_option' => $request->input('redirect_option')]);
 
         if ($user->save()) {
             // Redirect to the user page
@@ -352,10 +354,13 @@ class UsersController extends Controller
         if ($user = User::find($id)) {
 
             $this->authorize('delete', $user);
+            if (auth()->user()->can('canEditAuthFields', $user) && auth()->user()->can('editableOnDemo')) {
 
-            if ($user->delete()) {
-                return redirect()->route('users.index')->with('success', trans('admin/users/message.success.delete'));
+                if ($user->delete()) {
+                    return redirect()->route('users.index')->with('success', trans('admin/users/message.success.delete'));
+                }
             }
+            return redirect()->route('users.index')->with('error', trans('admin/users/message.cannot_delete'));
         }
         return redirect()->route('users.index')->with('error', trans('admin/users/message.user_not_found'));
