+++ resources/views/users/view.blade.php
@@ -206,12 +206,12 @@
                 <div class="col-md-12" style="padding-top: 5px;">
 
                 @if($user->allAssignedCount() != '0') 
-                  <a href="{{ route('users.print', $user->id) }}" style="width: 100%;" class="btn btn-sm btn-primary btn-social hidden-print" target="_blank" rel="noopener">
+                  <a href="{{ route('users.print', $user->id) }}" style="width: 100%;" class="btn btn-sm btn-theme btn-social hidden-print" target="_blank" rel="noopener">
                       <x-icon type="print" />
                       {{ trans('admin/users/general.print_assigned') }}
                   </a>
                   @else
-                  <button style="width: 100%;" class="btn btn-sm btn-primary btn-social hidden-print" rel="noopener" disabled title="{{ trans('admin/users/message.user_has_no_assets_assigned') }}">
+                  <button style="width: 100%;" class="btn btn-sm btn-theme btn-social hidden-print" rel="noopener" disabled title="{{ trans('admin/users/message.user_has_no_assets_assigned') }}">
                       <x-icon type="print" />
                       {{ trans('admin/users/general.print_assigned') }}</button>
                 @endif
@@ -223,7 +223,7 @@
                   @if(!empty($user->email) && ($user->allAssignedCount() != '0'))
                     <form action="{{ route('users.email',['userId'=> $user->id]) }}" method="POST">
                       {{ csrf_field() }}
-                      <button class="btn-block btn btn-sm btn-primary btn-social hidden-print" rel="noopener">
+                      <button class="btn-block btn btn-sm btn-theme btn-social hidden-print" rel="noopener">
                           <x-icon type="email" />
                           {{ trans('admin/users/general.email_assigned') }}
                       </button>
@@ -276,26 +276,32 @@
             @can('update', $user)
                   @if ($user->deleted_at=='')
                     <div class="col-md-12" style="padding-top: 30px;">
-                        @if ($user->isDeletable())
+                        @if (($user->isDeletable()) && ($user->id!=auth()->user()->id))
                             <a href="" class="delete-asset btn-block btn btn-sm btn-danger btn-social hidden-print" data-toggle="modal" data-title="{{ trans('general.delete') }}" data-content="{{ trans('general.sure_to_delete_var', ['item' => $user->display_name]) }}" data-icon="fa-trash" data-target="#dataConfirmModal" onClick="return false;" >
                                 <x-icon type="delete" />
                                 {{ trans('button.delete')}}
                             </a>
-                            @else
-                            <button class="btn-block btn btn-sm btn-danger btn-social hidden-print disabled">
+                        @elseif ($user->id == auth()->user()->id)
+                            <button class="btn-block btn btn-sm btn-danger btn-social hidden-print disabled" data-tooltip="true" data-title="{{ trans('tooltips.disabled_assoc.user_self') }}">
+                                <x-icon type="delete" />
+                                {{ trans('button.delete')}}
+                            </button>
+                        @else
+                            <button class="btn-block btn btn-sm btn-danger btn-social hidden-print disabled" data-tooltip="true" data-title="{{ trans('tooltips.disabled_assoc.user') }}">
                                 <x-icon type="delete" />
                                 {{ trans('button.delete')}}
                             </button>
                         @endif
                     </div>
                     <div class="col-md-12" style="padding-top: 5px;">
+
                       <form action="{{ route('users/bulkedit') }}" method="POST">
                         <!-- CSRF Token -->
                         <input type="hidden" name="_token" value="{{ csrf_token() }}" />
                         <input type="hidden" name="bulk_actions" value="delete" />
 
                         <input type="hidden" name="ids[{{ $user->id }}]" value="{{ $user->id }}" />
-                        <button class="btn btn-block btn-sm btn-danger btn-social hidden-print">
+                        <button class="btn btn-block btn-sm btn-danger btn-social hidden-print" data-tooltip="true" data-title="{{ trans('tooltips.checkin_all.user') }}">
                             <x-icon type="checkin-and-delete" />
                             {{ trans('button.checkin_and_delete') }}
                         </button>
@@ -845,8 +851,6 @@
 
             @include('partials.asset-bulk-actions')
 
-            <div class="table table-responsive">
-
             <table
                     data-columns="{{ \App\Presenters\AssetPresenter::dataTableLayout() }}"
                     data-show-columns-search="true"
@@ -867,13 +871,10 @@
                 "ignoreColumn": ["actions","image","change","checkbox","checkincheckout","icon"]
                 }'>
             </table>
-          </div>
         </div><!-- /asset -->
 
         <div class="tab-pane" id="licenses">
 
-
-          <div class="table-responsive">
             <table
                     data-cookie-id-table="userLicenseTable"
                     data-id-table="userLicenseTable"
@@ -929,7 +930,6 @@
                 @endforeach
               </tbody>
           </table>
-          </div>
         </div><!-- /licenses-tab -->
 
         <div class="tab-pane" id="accessories">
@@ -988,11 +988,9 @@
                   @endforeach
               </tbody>-->
             </table>
-          </div>
         </div><!-- /accessories-tab -->
 
         <div class="tab-pane" id="consumables">
-          <div class="table-responsive">
             <table
                     data-cookie-id-table="userConsumableTable"
                     data-id-table="userConsumableTable"
@@ -1027,7 +1025,6 @@
                 @endforeach
               </tbody>
           </table>
-          </div>
         </div><!-- /consumables-tab -->
 
         <div class="tab-pane" id="files">
@@ -1069,9 +1066,6 @@
           </div><!-- /eulas-tab -->
 
         <div class="tab-pane" id="history">
-          <div class="table-responsive">
-
-
               <table
                       data-columns="{{ \App\Presenters\HistoryPresenter::dataTableLayout() }}"
                       class="table table-striped snipe-table"
@@ -1086,8 +1080,6 @@
                      }'
                       data-url="{{ route('api.activity.index', ['item_id' => $user->id, 'item_type' => User::class]) }}">
               </table>
-
-          </div>
         </div><!-- /.tab-pane -->
 
         <div class="tab-pane" id="managed-locations">
@@ -1118,7 +1110,6 @@
           <div class="tab-pane" id="managed-users">
 
               @include('partials.users-bulk-actions')
-
 
               <table
                       data-columns="{{ \App\Presenters\UserPresenter::dataTableLayout() }}"
